refactor: consolidate authentication logic in useAuthMethods by creating a unified handleAuthAction function for login and registration, improving error handling with customizable messages; remove unused api service and update imports in auth and teaService to utilize axios directly
