+++ src/hooks/useAuthMethods.js
@@ -2,37 +2,27 @@ import { authService } from '../services/auth';
 import { setTokens, removeTokens } from '../utils/tokenStorage';
 
 export const useAuthMethods = (setUser, setIsLoading, setError) => {
-  const login = async (credentials) => {
+  const handleAuthAction = async (action, data, errorMessage) => {
     setIsLoading(true);
     setError(null);
     try {
-      const { data } = await authService.login(credentials);
-      setTokens(data.token, data.refreshToken);
-      setUser(data.user);
-      return data.user;
+      const { data: response } = await action(data);
+      setTokens(response.token, response.refreshToken);
+      setUser(response.user);
+      return response.user;
     } catch (err) {
-      setError(err.response?.data?.message || 'Login failed');
+      setError(err.response?.data?.message || errorMessage);
       throw err;
     } finally {
       setIsLoading(false);
     }
   };
 
-  const register = async (userData) => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const { data } = await authService.register(userData);
-      setTokens(data.token, data.refreshToken);
-      setUser(data.user);
-      return data.user;
-    } catch (err) {
-      setError(err.response?.data?.message || 'Registration failed');
-      throw err;
-    } finally {
-      setIsLoading(false);
-    }
-  };
+  const login = (credentials) => 
+    handleAuthAction(authService.login, credentials, 'Login failed');
+
+  const register = (userData) => 
+    handleAuthAction(authService.register, userData, 'Registration failed');
 
   const logout = () => {
     removeTokens();
